ON-46204 # Added changelog automation from a changelog-entries directory
