+++ changelog-entries/changelog-automation.md
@@ -1,0 +1,3 @@
+### Added
+
+- changelog automation from a `changelog-entries` directory
