+++ src/terminal/executeCommand.ts
