+++ src/terminal/wrapWithLoading.ts
